requirements.txt
asgiref==3.2.10
Django==3.0.8
django-appconf==1.0.4
django-ckeditor==5.9.0
django-image-cropping==1.4.0
django-js-asset==1.2.2
easy-thumbnails==2.7
install==1.3.3
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
